foods_api/serializers.py
from rest_framework import serializers
from .models import Food

class FoodSerializer(serializers.ModelSerializer):
    class Meta:
        model = Food
        fields = ('id', 'name', 'calories', 'image', 'created_at', 'updated_at', 'linked_users')
